feature/EAO-154 - Create service to get the actions of a Blocking
