+++ .github/workflows/build-deploy.yml
@@ -39,7 +39,7 @@ jobs:
           gradle wrapper
           ./gradlew wrapper --gradle-version=8.5
           ./gradlew wrapper --version
-          # ./gradlew spotlessCheck
+          ./gradlew spotlessCheck
           ./gradlew build -x spotlessGroovyGradle -x checkstyleMain -x spotbugsMain -x compileTestJava -x test -x cucumber -x checkstyleTest -x asciidoctorPdf -x asciidoctor -x buildAsciidoc -x spotbugsTest
           # ./gradlew sonar \
           # -Dsonar.host.url=$ENTRE_AMIGOS_SONAR_HOST_URL \
@@ -48,7 +48,6 @@ jobs:
           # -Dsonar.organization=$ENTRE_AMIGOS_SONAR_ORGANIZATION_KEY \
           # -Dsonar.qualitygate.wait=true \
           # --info
-
       - name: Prepare Test Server
         env:
           FINERACT_DEFAULT_TENANTDB_UID: postgres
@@ -66,6 +65,7 @@ jobs:
           RUN_ATF_PROCESS: true
 
         run: |
+          ./gradlew --stop
           ./gradlew dropPGDB -PdbName=$FINERACT_DEFAULT_TENANTDB_NAME -PdbUserName=$FINERACT_HIKARI_USERNAME -PdbUserPassword=$FINERACT_HIKARI_PASSWORD -PjdbcURL=$FINERACT_HIKARI_JDBC_URL
           ./gradlew createPGDB -PdbName=$FINERACT_DEFAULT_TENANTDB_NAME -PdbUserName=$FINERACT_HIKARI_USERNAME -PdbUserPassword=$FINERACT_HIKARI_PASSWORD -PjdbcURL=$FINERACT_HIKARI_JDBC_URL
           ./fineract-qa/pipetools/restore_dev_db.sh
@@ -92,7 +92,7 @@ jobs:
 
           echo "###############################################################################################"
           echo "#"
-          echo "# ⏳ Will run the collections now... Hold your breath and fingers crossed!"
+          echo "# ⏳ Will run the collections now... Hold your breath (not too long, please) and fingers crossed!"
           echo "#"
           for collection in fineract-qa/collections/*.postman_collection.json; do
             collection_name=$(basename "$collection" .postman_collection.json)
