+++ fineract-provider/src/main/java/org/apache/fineract/custom/portfolio/blockaccounts/service/impl/LoanAccountBlockReadPlatformServiceImpl.java
@@ -20,14 +20,21 @@ package org.apache.fineract.custom.portfolio.blockaccounts.service.impl;
  */
 import jakarta.transaction.Transactional;
 import jakarta.ws.rs.NotFoundException;
+import java.time.LocalDate;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.fineract.custom.portfolio.blockaccounts.data.LoanAccountBlockComponentEnum;
 import org.apache.fineract.custom.portfolio.blockaccounts.data.LoanAccountBlockDTO;
+import org.apache.fineract.custom.portfolio.blockaccounts.data.LoanAccountBlockData;
 import org.apache.fineract.custom.portfolio.blockaccounts.domain.LoanAccountBlock;
 import org.apache.fineract.custom.portfolio.blockaccounts.domain.LoanAccountBlockRepository;
 import org.apache.fineract.custom.portfolio.blockaccounts.mapper.LoanAccountBlockMapper;
 import org.apache.fineract.custom.portfolio.blockaccounts.service.LoanAccountBlockReadPlatformService;
+import org.apache.fineract.infrastructure.core.service.DateUtils;
 import org.springframework.stereotype.Service;
 
 @Slf4j
@@ -46,5 +53,78 @@ public class LoanAccountBlockReadPlatformServiceImpl implements LoanAccountBlock
             throw new NotFoundException(String.valueOf(loanId));
         }
         return loanAccountBlockMapper.toDto(loanAccountBlock.get());
+    }
+
+    @Override
+    public LoanAccountBlockData checkBlockAccountComponents(Long loanId, LocalDate givenDate) {
+        List<LoanAccountBlockComponentEnum> blockedComponents = new ArrayList<>();
+
+        Optional<LoanAccountBlock> loanAccountBlockOpt = loanAccountBlockRepository.retrieveByLoanIdAndStatusActive(loanId);
+
+        if (loanAccountBlockOpt.isPresent()) {
+            LoanAccountBlock loanAccountBlock = loanAccountBlockOpt.get();
+
+            if ("CASTIGO".equalsIgnoreCase(loanAccountBlock.getBlockingReasonSetting().getNameOfReason())
+                    && !DateUtils.getLocalDateOfTenant().isBefore(loanAccountBlock.getApplicationDate())) {
+
+                if (Objects.nonNull(loanAccountBlock.getAccelerate()) && loanAccountBlock.getAccelerate()) {
+                    blockedComponents.add(LoanAccountBlockComponentEnum.ACCELERATE);
+                }
+                if (Objects.nonNull(loanAccountBlock.getFreezeCurrentInterest()) && loanAccountBlock.getFreezeCurrentInterest()) {
+                    blockedComponents.add(LoanAccountBlockComponentEnum.FREEZE_INTEREST);
+                }
+                if (Objects.nonNull(loanAccountBlock.getFreezeInterestArrears()) && loanAccountBlock.getFreezeInterestArrears()) {
+                    blockedComponents.add(LoanAccountBlockComponentEnum.FREEZE_MORA);
+                }
+                if (Objects.nonNull(loanAccountBlock.getFreezeLifeInsurance()) && loanAccountBlock.getFreezeLifeInsurance()) {
+                    blockedComponents.add(LoanAccountBlockComponentEnum.FREEZE_LIFE_INSURANCE);
+                }
+                if (Objects.nonNull(loanAccountBlock.getFreezeMypime()) && loanAccountBlock.getFreezeMypime()) {
+                    blockedComponents.add(LoanAccountBlockComponentEnum.FREEZE_MIPYME);
+                }
+
+                if (blockedComponents.isEmpty()) {
+                    blockedComponents.add(LoanAccountBlockComponentEnum.BLOCK_DISBURSAL);
+                }
+            }
+        }
+
+        return LoanAccountBlockData.builder().loanId(loanId).providedDate(givenDate).loanAccountBlockComponentEnumList(blockedComponents)
+                .build();
+    }
+
+    @Override
+    public boolean containsBlockAccountDisbursal(Long loanId, LocalDate givenDate) {
+        return containsBlockAccount(loanId, givenDate, LoanAccountBlockComponentEnum.BLOCK_DISBURSAL);
+    }
+
+    @Override
+    public boolean containsBlockAccountAccelerate(Long loanId, LocalDate givenDate) {
+        return containsBlockAccount(loanId, givenDate, LoanAccountBlockComponentEnum.ACCELERATE);
+    }
+
+    @Override
+    public boolean containsBlockAccountFreezeInterest(Long loanId, LocalDate givenDate) {
+        return containsBlockAccount(loanId, givenDate, LoanAccountBlockComponentEnum.FREEZE_INTEREST);
+    }
+
+    @Override
+    public boolean containsBlockAccountFreezeMora(Long loanId, LocalDate givenDate) {
+        return containsBlockAccount(loanId, givenDate, LoanAccountBlockComponentEnum.FREEZE_MORA);
+    }
+
+    @Override
+    public boolean containsBlockAccountFreezeLifeInsurance(Long loanId, LocalDate givenDate) {
+        return containsBlockAccount(loanId, givenDate, LoanAccountBlockComponentEnum.FREEZE_LIFE_INSURANCE);
+    }
+
+    @Override
+    public boolean containsBlockAccountFreezeMipyme(Long loanId, LocalDate givenDate) {
+        return containsBlockAccount(loanId, givenDate, LoanAccountBlockComponentEnum.FREEZE_MIPYME);
+    }
+
+    private boolean containsBlockAccount(Long loanId, LocalDate givenDate, LoanAccountBlockComponentEnum blockComponentEnum) {
+        return checkBlockAccountComponents(loanId, givenDate).getLoanAccountBlockComponentEnumList().stream()
+                .anyMatch(disb -> disb.equals(blockComponentEnum));
     }
 }
