+++ fineract-provider/src/main/resources/db/changelog/tenant/parts/20250909_EA_86_create_table_loan_block_accounts.xml
@@ -109,4 +109,14 @@
         </modifySql>
     </changeSet>
 
+    <changeSet id="7" author="fineract">
+        <sql>
+            <![CDATA[
+                INSERT INTO m_blocking_reason_setting
+                (priority, description, name_of_reason, "level", createdby_id, created_date, lastmodifiedby_id, lastmodified_date, is_enabled, disabled_on_date, enabled_on_date, disabled_by, enabled_by, affects_client_level, affects_credit_level)
+                VALUES(9, 'CASTIGO', 'CASTIGO', 'CREDIT', NULL, '2025-01-23 14:14:26.262', NULL, '2025-01-23 14:14:26.262', 1, '2025-02-28 00:00:00.000', '2025-04-22 00:00:00.000', 1, 12, NULL, NULL);
+            ]]>
+        </sql>
+    </changeSet>
+
 </databaseChangeLog>
