+++ fineract-qa/collections/EA Loans AccountBlock.postman_collection.json
@@ -1,10 +1,9 @@
 {
     "info": {
-        "_postman_id": "9e5815b0-0299-4b3b-ba38-ee78f9fc35d6",
-        "name": "EA Block Loan Accounts",
+        "_postman_id": "ed7899c1-51d5-490e-bcf1-48334946c8cc",
+        "name": "EA Loans AccountBlock",
         "schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json",
-        "_exporter_id": "34930034",
-        "_collection_link": "https://jacobo-vargas.postman.co/workspace/69da17f8-6926-42b5-8ba0-23bc6d424a86/collection/34930034-9e5815b0-0299-4b3b-ba38-ee78f9fc35d6?action=share&source=collection_link&creator=34930034"
+        "_exporter_id": "26949565"
     },
     "item": [
         {
@@ -375,7 +374,7 @@
             "response": []
         },
         {
-            "name": "Block Account Fail Case",
+            "name": "Test Block Account Fail Case",
             "event": [
                 {
                     "listen": "test",
@@ -436,6 +435,85 @@
             "response": []
         },
         {
+            "name": "Get Blocking Reason Castigo",
+            "event": [
+                {
+                    "listen": "test",
+                    "script": {
+                        "exec": [
+                            "pm.response.to.have.status(200);\r",
+                            "pm.expect(pm.response.text()).to.not.equals( null );\r",
+                            "var jsonData = JSON.parse(pm.response.text());\r",
+                            "\r",
+                            "// Encontra o objeto com o \"nameOfReason\" igual a \"CASTIGO\"\r",
+                            "const castigoRecord = jsonData.find(record => record.nameOfReason === \"CASTIGO\");\r",
+                            "\r",
+                            "// Se o registro for encontrado, salva o 'id' em uma variável de ambiente\r",
+                            "if (castigoRecord) {\r",
+                            "    pm.environment.set(\"castigoId\", castigoRecord.id);\r",
+                            "    console.log(\"ID do 'CASTIGO' salvo na variável 'id_castigo': \" + castigoRecord.id);\r",
+                            "} else {\r",
+                            "    console.log(\"Registro 'CASTIGO' não encontrado na resposta.\");\r",
+                            "}"
+                        ],
+                        "type": "text/javascript",
+                        "packages": {}
+                    }
+                },
+                {
+                    "listen": "prerequest",
+                    "script": {
+                        "exec": [
+                            ""
+                        ],
+                        "type": "text/javascript",
+                        "packages": {}
+                    }
+                }
+            ],
+            "protocolProfileBehavior": {
+                "disableBodyPruning": true
+            },
+            "request": {
+                "method": "GET",
+                "header": [
+                    {
+                        "key": "Authorization",
+                        "value": "{{authorization_token}}",
+                        "type": "text"
+                    },
+                    {
+                        "key": "Fineract-Platform-TenantId",
+                        "value": "default",
+                        "type": "text"
+                    }
+                ],
+                "body": {
+                    "mode": "raw",
+                    "raw": "{\r\n    \"transactionDate\": \"18/07/2025\",\r\n    \"transactionAmount\": 200000,\r\n    \"honorariosAmount\": 0,\r\n    \"externalId\": \"\",\r\n    \"paymentTypeId\": 1,\r\n    \"note\": \"\",\r\n    \"channelName\": \"03 - Oficina BCS\",\r\n    \"repaymentBankId\": \"\",\r\n    \"pointOfSalesCode\": \"\",\r\n    \"reduceInstallmentAmount\": true,\r\n    \"dateFormat\": \"dd/MM/yyyy\",\r\n    \"locale\": \"es\"\r\n}",
+                    "options": {
+                        "raw": {
+                            "language": "json"
+                        }
+                    }
+                },
+                "url": {
+                    "raw": "{{fineract-backend-server}}/fineract-provider/api/v1/blockSettings/retrieveAllBlockingReasons",
+                    "host": [
+                        "{{fineract-backend-server}}"
+                    ],
+                    "path": [
+                        "fineract-provider",
+                        "api",
+                        "v1",
+                        "blockSettings",
+                        "retrieveAllBlockingReasons"
+                    ]
+                }
+            },
+            "response": []
+        },
+        {
             "name": "Block Account",
             "event": [
                 {
@@ -473,7 +551,7 @@
                 ],
                 "body": {
                     "mode": "raw",
-                    "raw": "{\r\n  \"applicationDate\": \"11/07/2025\",\r\n  \"blockingReasonId\": 24,\r\n  \"accelerate\": false,\r\n  \"freezeCurrentInterest\": true,\r\n  \"freezeInterestArrears\": true,\r\n  \"freezeLifeInsurance\": false,\r\n  \"freezeMypime\": false,\r\n  \"active\": true,\r\n  \"locale\": \"es\",\r\n  \"dateFormat\": \"dd/MM/yyyy\"\r\n}\r\n",
+                    "raw": "{\r\n  \"applicationDate\": \"11/07/2025\",\r\n  \"blockingReasonId\": {{castigoId}},\r\n  \"accelerate\": false,\r\n  \"freezeCurrentInterest\": true,\r\n  \"freezeInterestArrears\": true,\r\n  \"freezeLifeInsurance\": false,\r\n  \"freezeMypime\": false,\r\n  \"active\": true,\r\n  \"locale\": \"es\",\r\n  \"dateFormat\": \"dd/MM/yyyy\"\r\n}\r\n",
                     "options": {
                         "raw": {
                             "language": "json"
@@ -497,7 +575,7 @@
             "response": []
         },
         {
-            "name": "Block Account Already exists",
+            "name": "Test Block Account Already exists",
             "event": [
                 {
                     "listen": "test",
